requirements.txt
astroid==2.4.1
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.42.0
django-bootstrap4==2.0.1
django-crispy-forms==1.9.1
django-heroku==0.3.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.2
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
